@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@react-pdf/renderer": "^2.2.0",
+    "@react-pdf/renderer": "^3.1.14",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.26.1",
     "bootstrap": "^5.1.3",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.2.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-github-calendar": "^3.2.2",
     "react-icons": "^4.8.0",
     "react-parallax-tilt": "^1.7.42",
-    "react-pdf": "^5.7.1",
-    "react-router-dom": "^6.2.2",
+    "react-pdf": "^7.7.1",
+    "react-router-dom": "^6.22.0",
     "react-scripts": "5.0.0",
     "react-tsparticles": "^1.42.2",
     "sharp": "^0.33.5",
