@@ -12,7 +12,7 @@ function Home() {
       <div className="main-section">
         <div className="grid-pattern"></div>
         <Container className="home-content">
-          <Row>
+          <Row className="align-items-center">
             <Col md={7} className="home-header">
               <h1 style={{ paddingBottom: 15 }} className="heading">
                 Hi There!{" "}
@@ -26,17 +26,19 @@ function Home() {
                 <strong className="main-name"> MARKO PETRINA</strong>
               </h1>
 
-              <div style={{ padding: "50px 50px 25px 50px", textAlign: "left" }}>
+              <div style={{ padding: 50, textAlign: "left" }}>
                 <Type />
               </div>
             </Col>
-            <Col md={3} style={{ paddingBottom: 20, paddingTop: 0, marginTop: "-20px" }}>  
+            <Col md={5} xs={12} className="avatar-column">
               <TiltAvatar />
             </Col>
           </Row>
         </Container>
       </div>
-      <Home2 />
+      <div id="home2">
+        <Home2 />
+      </div>
     </section>
   );
 }
